+++ src/app/internal/type-generator.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, Renderer } from "@angular/core";
+import { Component, OnInit, Renderer2 } from "@angular/core";
 import { NgForm } from "@angular/forms";
 // types